Rename stringifyVariables.ts to variables.ts
