+++ packages/core/src/utils/index.ts
@@ -2,7 +2,7 @@ export * from './error';
 export * from './request';
 export * from './result';
 export * from './typenames';
-export * from './stringifyVariables';
+export * from './variables';
 export * from './maskTypename';
 export * from './streamUtils';
 export * from './operation';
