+++ packages/core/src/utils/variables.ts
